requirements.txt
evaluate>=0.4.0
google-generativeai>=0.8.3
numpy>=1.26.4
openai>=1.0.0
python-dotenv>=1.0.1